* feature/gridtools-storage: (173 commits)
  Reuse array_initializer for native implementation
  Move test_array* from tests/util to tests/array
  Fix running test_trans_serial on Cray with MPI=ON
  Require ecbuild 2.5, eckit 0.15, fckit 0.3, transi 0.4
  Version 0.9.0
  Fix warnings in documentation code
  add real32 for rank4 halo exchange
  Missing include
  Fix warnings
  Fix bug in order_independent_sum (idx out of bounds)
  Fix compilation for new Array implementation
  ATLAS-73 Rename unit-test target
  ATLAS-73 Fortran interface
  ATLAS-73 Fixes and unit-test
  ATLAS-73 Expose trans function to compute derivatives from spectral (untested)
  ATLAS-77 Spectral norms, Atlas interface
  Flush Log buffers when atlas_finalize() is called
  Fix sign comparison warnings
  Remove pragma's
  Fix sign comparison warnings
  ...

Conflicts:
	src/atlas/mesh/Connectivity.cc
	src/atlas/mesh/Connectivity.h

+++ src/atlas/array/ArraySpec.cc
@@ -76,7 +76,7 @@ ArraySpec::ArraySpec( const ArrayShape& shape, const ArrayStrides& strides, cons
     strides_[j] = strides[j];
     layout_[j]  = layout[j];
     size_ *= shape_[j];
-    if( layout_[j] != j ) default_layout_ = false;
+    if( layout_[j] != size_t(j) ) default_layout_ = false;
   }
   contiguous_ = (size_ == shape_[0]*strides_[0] ? true : false);
 }
